Estimate the mixture proportions and component densities using EM algorithm

Uses EM algorithm to estimate the mixture proportions and the component densities. The output is an object of class "lcdmix" which contains mixture proportions at each observation and all the information of the estimated component densities.

Usage

EMmixlcd( x, k = 2, y, props, epsratio=10^-6, max.iter=50,
            epstheta=10^-8, verbose=-1 )

Arguments

x

Data in \(R^d\), in the form of an \(n \times d\) numeric matrix

k

The number of components, equals 2 by default

y

An \(n \times k\) numeric matrix giving the starting values for the EM algorithm. If none given, a hierachical Gaussian clustering model is used. To reduce the computational burden while allowing sufficient flexibility for the EM algorithm, it is recommended to leave this argument unspecified.

props

Vector of length \(k\) containing the starting value of proportions. If none given, a hierachical Gaussian clustering model is used. To reduce the computational burden while allowing sufficient flexibility for the EM algorithm, it is recommended to leave this argument unspecified.

epsratio

EM algorithm will terminate if the increase in the proportion of the likelihood is less than this specified ratio. Default value is \(10^{-6}\).

max.iter

The maximum number of iterations for the EM algorithm

epstheta

\(epstheta/n\) is the thresold of the weight below which data point is discarded from the cluster. This quantity is introduced to increase the computational efficiency and stability.

verbose

-1: (default) prints nothing

0: prints warning messages

\(>0\): prints summary information every \(n\) iterations

Details

An introduction to the Em algorithm can be found in McLachlan and Krishnan (1997). Briefly, given the current estimates of the mixture proportions and component densities, we first update the estimates of the mixture prroportions. We then update the estimates of the component densities by using mlelcd. In fact, the incorporation of the weights in the maximization process in mlelcd presents no additional complication.

In our case, because of the computational intensity of the method, we first cluster the points according to ta hierarchical Gaussian clustering model and then iterate the EM algorithm until the increase in the proportion of the likelihood is less than a pre-specified quantity at each step.

More technical details can be found in Cule, Samworth and Stewart(2010)

Value

An object of class "lcdmix", with the following components:

x

Data copied from input (may be reordered)

logf

An \(n \times k\) maxtrix of the log of the maximum likelihood estimate, evaluated at the observation points for each component.

props

Vector containing the estimated proportions of components

niter

Number of iterations of the EM algorithm

lcdloglik

The log-likelihood after the final iteration

References

Cule, M. L., Samworth, R. J., and Stewart, M. I. (2010) Maximum likelihood estimation of a log-concave density, Journal of the Royal Statistical Society, Series B, 72(5) p.545-607.

McLachlan, G. J. and Krishnan, T. (1997) The EM Algorithm and Extensions, New York: Wiley.

Author

Yining Chen

Madeleine Cule

Robert B. Gramacy

Richard Samworth

See also

Examples

##Simple bivariate normal data
  set.seed( 1 )
  n = 15
  d = 2
  props=c( 0.6, 0.4 )
  shift=2
  x <- matrix( rnorm( n*d ), ncol = d )
  shiftvec <- ifelse( runif( n ) > props[ 1 ], 0, shift )
  x[,1] <- x[,1] + shiftvec
  EMmixlcd( x, k = 2, max.iter = 2)
#> $x
#>             [,1]        [,2]
#>  [1,] -0.6264538 -0.04493361
#>  [2,]  2.1836433 -0.01619026
#>  [3,]  1.1643714  0.94383621
#>  [4,]  3.5952808  0.82122120
#>  [5,]  0.3295078  0.59390132
#>  [6,]  1.1795316  0.91897737
#>  [7,]  2.4874291  0.78213630
#>  [8,]  0.7383247  0.07456498
#>  [9,]  2.5757814 -1.98935170
#> [10,] -0.3053884  0.61982575
#> [11,]  3.5117812 -0.05612874
#> [12,]  0.3898432 -0.15579551
#> [13,]  1.3787594 -1.47075238
#> [14,] -0.2146999 -0.47815006
#> [15,]  3.1249309  0.41794156
#> 
#> $logf
#>             [,1]       [,2]
#>  [1,] -0.9406511       -Inf
#>  [2,]       -Inf -1.0749647
#>  [3,]  0.2729687       -Inf
#>  [4,]       -Inf -0.5484670
#>  [5,] -0.3155728       -Inf
#>  [6,]  0.2900564       -Inf
#>  [7,]       -Inf -0.6120231
#>  [8,]  0.1142542       -Inf
#>  [9,]       -Inf -2.1764194
#> [10,] -0.8230307       -Inf
#> [11,]       -Inf -1.0478670
#> [12,] -0.1141248       -Inf
#> [13,]       -Inf -1.9278556
#> [14,] -0.5261604       -Inf
#> [15,]       -Inf -0.7941805
#> 
#> $props
#> [1] 0.5333333 0.4666667
#> 
#> $niter
#> [1] 3
#> 
#> $lcdloglik
#> [1] -20.58789
#> 
#> attr(,"class")
#> [1] "lcdmix"