Evaluation of a log-concave maximum likelihood estimator at a point

This function evaluates the density function of a log-concave maximum likelihood estimator at a point or points.

Usage

dlcd(x,lcd, uselog=FALSE, eps=10^-10)

Arguments

x

Point (or matrix of points) at which the maximum likelihood estimator should be evaluated

lcd

Object of class "LogConcDEAD" (typically output from mlelcd)

uselog

Scalar logical: should the estimator should be calculated on the log scale?

eps

Tolerance for numerical stability

Details

A log-concave maximum likelihood estimate \(\hat{f}_n\) is satisfies \(\log \hat{f}_n = \bar{h}_y\) for some \(y \in R^n\), where $$\bar{h}_y(x) = \inf \lbrace h(x) \colon h \textrm{ concave }, h(x_i) \geq y_i \textrm{ for } i = 1, \ldots, n \rbrace.$$

Functions of this form may equivalently be specified by dividing \(C_n\), the convex hull of the data into simplices \(C_j\) for \(j \in J\) (triangles in 2d, tetrahedra in 3d etc), and setting $$f(x) = \exp\{b_j^T x - \beta_j\}$$ for \(x \in C_j\), and \(f(x) = 0\) for \(x \notin C_n\). The estimated density is zero outside the convex hull of the data.

The estimate may therefore be evaluated by finding the appropriate simplex \(C_j\), then evaluating \(\exp\{b_j^T x - \beta_j\}\) (if \(x \notin C_n\), set \(f(x) = 0\)).

For examples, see mlelcd.

Value

A vector of maximum likelihood estimate (or log maximum likelihood estimate) values, as evaluated at the points x.

Author

Madeleine Cule

Robert Gramacy

Richard Samworth

See also