Construct an object of class LogConcDEAD

A function to construct an object of class LogConcDEAD from a dataset (given as a matrix) and the value of the log maximum likelihood estimator at datapoints.

Usage

getinfolcd(x, y, w = rep(1/length(y), length(y)), chtol = 10^-6, 
  MinSigma = NA, NumberOfEvaluations = NA)

Arguments

x

Data in \(R^d\), in the form of an \( n \times d\) numeric matrix

y

Value of log of maximum likelihood estimator at data points

w

Vector of weights \(w_i\) such that the computed estimator maximizes $$\sum_{i=1}^n w_i \log f(x_i)$$ subject to the restriction that \(f\) is log-concave. The default is \(\frac{1}{n}\) for all \(i\), which corresponds to i.i.d. observations.

chtol

Tolerance for computation of convex hull. Altering this is not recommended.

MinSigma

Real-valued scalar giving minimum value of the objective function

NumberOfEvaluations

Vector containing the number of steps, number of function evaluations, and number of subgradient evaluations. If the SolvOpt algorithm fails, the first component will be an error code \((<0)\)

Details

This function is used in mlelcd

Value

An object of class "LogConcDEAD", with the following components:

x

Data copied from input (may be reordered)

w

weights copied from input (may be reordered)

logMLE

vector of the log of the maximum likelihood estimate, evaluated at the observation points

NumberOfEvaluations

Vector containing the number of steps, number of function evaluations, and number of subgradient evaluations. If the SolvOpt algorithm fails, the first component will be an error code \((<0)\).

MinSigma

Real-valued scalar giving minimum value of the objective function

b

matrix (see Details)

beta

vector (see Details)

triang

matrix containing final triangulation of the convex hull of the data

verts

matrix containing details of triangulation for use in dlcd

vertsoffset

matrix containing details of triangulation for use in dlcd

chull

Vector containing vertices of faces of the convex hull of the data

outnorm

matrix where each row is an outward pointing normal vectors for the faces of the convex hull of the data. The number of vectors depends on the number of faces of the convex hull.

outoffset

matrix where each row is a point on a face of the convex hull of the data. The number of vectors depends on the number of faces of the convex hull.

Author

Madeleine Cule

Robert B. Gramacy

Richard Samworth

Yining Chen

See also